.travis.yml: re-enable email notifications
We previously disabled them because of issues #126 and #127, which would
make travis-ci always fail with the state of the codebase at that time.
Since the problem was already solved (even though #126 is still opened),
let's revoke the email notification prohibition. Note: we still disable
notifications for success builds in order to reduce spamming.

Reviewed-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>
Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>