mk: Add cleanup function for kw deploy
kw deploy creates a directory for centralizing all files that we need to
use for updating a kernel in a remote machine. We don't need to keep
these files in the hard disk for a long time, at the end of the deploy
we can remove it. This commit adds a function that cleans all temporary
files after a kernel deploys finish.

Closes #170

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>