statistics: Introduce the statistics feature process
This commit introduces a statistics option that is responsible to
provide kw usage summary; users can see this data based on four levels
of report granularity: day, week, month, and year. Notice that this
commit goal is to introduce the feature in a way that can be expanded in
the future.

Closes #104

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>