help: Break help function by sub-commands
A single help command for the entirety of kw does not scale, and it
becomes excessively long and complex to be useful. This commit
reorganizes the help information by keeping a single helper function
with a brief description of each option in the kw core and add helper
functions per feature with extra details on how to use it.

closes #180

Signed-off-by: alan-barzilay <alan.barzilay@gmail.com>
Reviewed-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>