src: kw_string: Add a file dedicated to handling string
Unfortunately, some of the shell parameter expansion for handling
strings are not obvious at first glance; in particular, those not used
to how bash works might not understand the syntax. For trying to make kw
code more readable, this commit starts a tiny library that comprises
some of the basic string manipulation currently used in the source.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>
Reviewed-by: Alan Barzilay <alan.barzilay@gmail.com>
Reviewed-by: Magali Lemes <magalilemes00@gmail.com>