src: remote: move ssh functionality to remote.sh
The current functions that handle the ssh functionality of kw are
located in the `vm.sh` file, it makes more sense for it to be in the
`remote.sh` file.  This commit simply moves the ssh functionality from
one file to the other making the necessary adjustments so the code
continues to execute as it did before.

This also moves all related ssh tests to the `remote_test.sh`, and
deletes `ssh_test.sh`.

Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@usp.br>
Reviewed-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>